CI: Run CI every day, not only Mon-Fri
We used to have the CI running only for work days to not bother
people on their weekends with reports of problems.

That in its turn postpones testing of changes done on Friday
till the next Monday.

But why not running all tests daily to get rid of possible delays?