Updated newlib allowlist
thread_local-order2.C : TLS is expected to FAIL on Baremetal applications
due to the missing operating system.

raw-string-7.c : Raw string literals are expected to FAIL on all
architectures due to a wrong conversion of tab character.

pr82274-1.c : GCC correctly detects integer overflow and aborts the program
as expected when the `ftrapv` flag is enabled.