Rectified inaccurate Argument in nSIM Wrapper Configuration.
With the integration of the QEMU testsuite, a change in the
configuration arguments for the QEMU wrapper was introduced.
The `-semihosting` flag, previously directly applied in the
wrapper, has been removed and is now provided as an argument
`-Wq,-semihosting`. This adjustment was necessary because
the QEMU Testsuite does not support enabling the flag
in its former fashion.

This commit addresses the issue where `-Wq,semihosting`
was improperly passed to the nSIM wrapper, causing a crash.

Signed-off-by: Luis Silva <luis.m.silva99@hotmail.com>