Explicitly set the environment version to exlude extra build errors
The purpose of this commit is to provide predictable behaviour for the build process. It is necessary because another project (https://github.com/foss-for-synopsys-dwc-arc-processors/linux) also depends on this CI. The environment version can be changed later as the project progresses, but don't forget to change the dependent projects too.

Signed-off-by: Evgeny Semenov <semenov@synopsys.com>